+++ clients/core/src/Sidebar/CourseSwitcher.tsx
@@ -17,7 +17,7 @@ import {
   useSidebar,
 } from '@/components/ui/sidebar'
 
-export function TeamSwitcher({
+export function CourseSwitcher({
   teams,
 }: {
   teams: {
