+++ clients/core/src/management/ManagementConsole.tsx
@@ -2,7 +2,7 @@ import { useKeycloak } from '@/keycloak/useKeycloak'
 import { useAuthStore } from '@/zustand/useAuthStore'
 import UnauthorizedPage from './components/UnauthorizedPage'
 import { SidebarProvider } from '@/components/ui/sidebar'
-import { AppSidebar } from '../Sidebar/app-sidebar'
+import { AppSidebar } from '../Sidebar/AppSidebar'
 import { WelcomePage } from './components/WelcomePage'
 import { LoadingPage } from '@/components/LoadingPage'
 import React from 'react'
